GTKUI/requirements.txt
PyGObject
gpiozero
pyserial
